refactor(timeline): extract TimelineContent row components
Move the five module-private row/entry components into components/ and
lift their shared lineage types into the existing types.ts. index.tsx
now only holds the container. Pure code motion; behaviour unchanged.
